refactor LegacyAccountBackup
